+++ src/scriptengine.cpp
@@ -353,14 +353,14 @@ void framework::scriptengine::run() {
   );
 
   struct playerwrapper {
-    int32_t index;
+    uint8_t index;
     const framework::statemanager &e;
 
     playerwrapper(input::event::player player, const framework::statemanager &state_manager)
-      : index(static_cast<int32_t>(player)), e(state_manager) {}
+      : index(static_cast<uint8_t>(player)), e(state_manager) {}
 
     bool on(std::variant<input::event::gamepad::button> type) {
-      return e.on(index, type);
+      return e.on(static_cast<uint8_t>(index), type);
     }
   };
