+++ src/statemanager.hpp
@@ -22,9 +22,9 @@ public:
 
   bool collides(std::shared_ptr<object> a, std::shared_ptr<object> b) const;
 
-  bool on(int player, const std::variant<input::event::gamepad::button> &type) const;
+  bool on(uint8_t player, const std::variant<input::event::gamepad::button> &type) const;
 
-  int8_t players() const;
+  uint8_t players() const;
 
 protected:
   virtual void on_keydown(const input::event::keyboard::key &event);
@@ -42,7 +42,7 @@ protected:
   virtual void on_endupdate();
 
 private:
-  std::unordered_map<int8_t, std::unordered_map<std::variant<input::event::gamepad::button>, bool>> _state;
+  std::unordered_map<uint8_t, std::unordered_map<std::variant<input::event::gamepad::button>, bool>> _state;
 
   std::unordered_map<std::pair<uint64_t, uint64_t>, bool, pairhash> _collision_mapping;
 };
